requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
pyyaml>=6.0

# Deep learning
torch>=2.0
torchvision>=0.15

# EEG / neuro
mne>=1.6

# Reconstruction metrics
scikit-image>=0.21       # SSIM
transformers>=4.35       # CLIP via HuggingFace
Pillow>=10.0

# Visualization
matplotlib>=3.7
seaborn>=0.13

# Notebooks
jupyter>=1.0
ipykernel>=6.0

# Optional: video generation (EEG2Video pipeline)
# diffusers>=0.24
# accelerate>=0.24
# einops>=0.7
